requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pandas==2.2.2
numpy==1.26.4
requests==2.31.0
beautifulsoup4==4.12.2
matplotlib==3.8.1
seaborn==0.12.2
duckdb==1.1.2
python-multipart==0.0.6
Pillow==10.1.0
openai
boto3==1.28.80
python-dotenv==1.0.0
